src/simulator/requirements.txt
azure-eventhub==5.10.1
pandas